refactor: 파일 구조 craco 방식 사용하여 수정
