+++ jsconfig.json
@@ -38,12 +38,20 @@
 			"@HowToUse/*": ["components/HowToUse/*"],
 			"@RoadMap": ["components/RoadMap/index"],
 			"@RoadMap/*": ["components/RoadMap/*"],
-			"@Competency": ["components/RoadMap/old/Competency/index"],
-			"@Competency/*": ["components/RoadMap/old/Competency/*"],
-			"@RoadMapContents": ["components/RoadMap/old/RoadMap/index"],
-			"@RoadMapContents/*": ["components/RoadMap/old/RoadMap/*"],
-			"@TotalRoadMap": ["components/RoadMap/old/TotalRoadMap/index"],
-			"@TotalRoadMap/*": ["components/RoadMap/old/TotalRoadMap/*"],
+			"@Table": ["components/RoadMap/Table/index"],
+			"@Table/*": ["components/RoadMap/Table/*"],
+			"@Competency": ["components/RoadMap/Competency/index"],
+			"@Competency/*": ["components/RoadMap/Competency/*"],
+			"@Course": ["components/RoadMap/Course/index"],
+			"@Course/*": ["components/RoadMap/Course/*"],
+			"@TotalRoadMap": ["components/RoadMap/TotalRoadMap/index"],
+			"@TotalRoadMap/*": ["components/RoadMap/TotalRoadMap/*"],
+			"@CompetencyOld": ["components/RoadMap/old/Competency/index"],
+			"@CompetencyOld/*": ["components/RoadMap/old/Competency/*"],
+			"@RoadMapContentsOld": ["components/RoadMap/old/RoadMap/index"],
+			"@RoadMapContentsOld/*": ["components/RoadMap/old/RoadMap/*"],
+			"@TotalRoadMapOld": ["components/RoadMap/old/TotalRoadMap/index"],
+			"@TotalRoadMapOld/*": ["components/RoadMap/old/TotalRoadMap/*"],
 			"@hooks": ["hooks/index"],
 			"@hooks/*": ["hooks/*"],
 			"@recoils": ["recoils/index"],
