+++ src/components/RoadMap/Competency/CompetencyTable.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import styled from 'styled-components';
-import CompetencyCell from './CompetencyCell';
-import { Color } from '../../styles/Color';
+import { CompetencyCell } from '@Competency';
+import { Color } from '@styles';
 
 const Container = styled.div`
 	width: 15%;
