+++ src/components/RoadMap/Competency/index.js
@@ -1,0 +1,2 @@
+export { default as CompetencyTable } from './CompetencyTable';
+export { default as CompetencyCell } from './CompetencyCell';
