+++ src/components/RoadMap/Course/CourseCell.jsx
@@ -1,10 +1,9 @@
 import { useState, useEffect, useRef } from 'react';
 import styled from 'styled-components';
 import { useRecoilValue, useSetRecoilState } from 'recoil';
-import { selectedCompetencyState, selectedMyTableContentsState } from '../../recoils/atoms';
-import CourseDetail from '../DetailContents/CourseDetail/CourseDetail';
-import { fadeIn } from '../../styles/Frames';
-import { Color } from '../../styles/Color';
+import { selectedCompetencyState, selectedMyTableContentsState } from '@recoils';
+import { CourseDetail } from '@CourseDetail';
+import { Color, fadeIn } from '@styles';
 
 const Button = styled.div`
 	width: 100%;
