+++ src/components/RoadMap/Course/CourseTable.jsx
@@ -1,8 +1,7 @@
 import React, { useRef } from 'react';
 import styled from 'styled-components';
-import SemesterCell from './SemesterCell';
-import CourseCell from './CourseCell';
-import { SemesterTable } from '../Common/Utils';
+import { SemesterCell, CourseCell } from '@Course';
+import { SemesterTable } from '@Common/Utils';
 
 const TableContainer = styled.div`
 	flex: 1;
