+++ src/components/RoadMap/Course/index.js
@@ -1,0 +1,3 @@
+export { default as CourseTable } from './CourseTable';
+export { default as CourseCell } from './CourseCell';
+export { default as SemesterCell } from './SemesterCell';
