+++ src/components/RoadMap/Table/RoadMapTable.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import CourseTable from './CourseTable';
-import CompetencyTable from './CompetencyTable';
+import { CourseTable } from '@Course';
+import { CompetencyTable } from '@Competency';
 
 const Container = styled.div`
 	height: 21rem;
