+++ src/components/RoadMap/Table/index.js
@@ -1,0 +1,3 @@
+export { default as CourseCreditTable } from './CourseCreditTable';
+export { default as MyMapTable } from './MyMapTable';
+export { default as RoadMapTable } from './RoadMapTable';
