+++ src/components/RoadMap/index.js
@@ -1,12 +1,7 @@
+export * from './Competency';
+export * from './Course';
+export * from './Table';
 export * from './TotalRoadMap';
 
 export { default as Main } from './Main';
-export { default as CompetencyCell } from './CompetencyCell';
-export { default as CompetencyTable } from './CompetencyTable';
-export { default as CourseCell } from './CourseCell';
-export { default as SemesterCell } from './SemesterCell';
-export { default as CourseCreditTable } from './CourseCreditTable';
-export { default as CourseTable } from './CourseTable';
-export { default as MyMapTable } from './MyMapTable';
-export { default as RoadMapTable } from './RoadMapTable';
-export { default as RoadMapContentContainer } from './RoadMapContentContainer';
+export { default as RoadMapContents } from './RoadMapContents';
